db1/requirements.txt
Django==2.0
supervisor
DateTime==4.3
Django==2.0
django-crispy-forms==1.7.2
djangorestframework==3.10.2
numpy==1.17.0
pandas==0.25.0
python-dateutil==2.8.0
six==1.12.0
sqlparse==0.3.0
zope.interface==4.6.0
requests
celery
python-googlegeocoder
virtualenv==16.7.2
pytz
django-cors-headers
django-rest-auth
django-allauth